it_ = deque(maxlen=r), etc. in comp(), renamed level_1 -> level_1_working, renamed level_1_working -> level_1_old
