+++ level_1_working.py
@@ -1,189 +1,205 @@
 from scipy import misc
+from collections import deque
 
 '''
 Level 1:
-Cross-comparison between consecutive pixels within horizontal scan line (row).
-Resulting difference patterns dPs: spans of pixels forming same-sign differences,
-and relative match patterns vPs: spans of pixels forming same-sign predictive value,
-overlap within each line of pixels.
+
+Cross-comparison between consecutive pixels within horizontal scan level (row).
+Resulting difference patterns dPs (spans of pixels forming same-sign differences)
+and relative match patterns vPs (spans of pixels forming same-sign predictive value)
+are redundant representations of each level of pixels.
+
+I don't like to pack arguments, this code is optimized for visibility rather than speed 
 '''
 
-def inc_rng(a, aV, aD, min_r, A, AV, AD, r, p_):
 
-    if r > min_r:  # A, AV, AD inc.to adjust for redundancy to patterns formed by prior comp:
-        A += a     # a: min m for inclusion into positive vP
-        AV += aV   # aV: min V for initial comp() recursion, AV: min V for higher recursions
+def pre_comp(typ, e_, A, r):  # pre-processing for comp recursion
 
-    if r > min_r-1:  # default range is shorter for d_[w]: redundant ds are smaller than ps
-        AD += aD     # aV: min |D| for comp() recursion over d_[w], AD: min |D| for recursion
+    A += a  # compensates for fv overlap redundancy
+    X = len(e_)
 
-    X = len(p_)
-    ip_ = p_  # to differentiate from new p_
+    olp, vP_, dP_ = 0, [], []   # olp is common for both:
+    vP = 0, 0, 0, 0, 0, [], []  # pri_s, I, D, V, rv, t_, olp_
+    dP = 0, 0, 0, 0, 0, [], []  # pri_sd, Id, Dd, Vd, rd, d_, dolp_
 
-    vP_, dP_ = [],[]  # r was incremented in higher-scope p_
-    pri_s, I, D, V, rv, olp, p_, olp_ = 0, 0, 0, 0, 0, 0, [], []  # tuple vP=0
-    pri_sd, Id, Dd, Vd, rd, dolp, d_, dolp_ = 0, 0, 0, 0, 0, 0, [], []  # tuple dP=0
+    if typ:  # comparison range incr within e_ = t_
 
-    for x in range(r+1, X):
+        r += 1  # comp range counter, recorded within Ps formed by re_comp
+        for x in range(r+1, X):
 
-        p, fd, fv = ip_[x]       # compared to a pixel at x-r-1:
-        pp, pfd, pfv = ip_[x-r]  # previously compared p(ignored), its fd, fv to next p
-        fv += pfv  # fv is summed over extended-comp range
-        fd += pfd  # fd is summed over extended-comp range
+            p, ifd, ifv = e_[x]  # ifd, ifv not used, directional pri_p accum only
+            pri_p, fd, fv = e_[x-r]  # for comparison of r-pixel-distant pixels:
 
-        pri_p, pri_fd, pri_fv = ip_[x-r-1]  # for comp(p, pri_p), pri_fd and pri_fv ignored
+            fd, fv, vP, dP, vP_, dP_, olp = \
+            re_comp(x, p, pri_p, fd, fv, vP, dP, vP_, dP_, olp, X, A, r)
 
-        pri_s, I, D, V, rv, p_, olp, olp_, pri_sd, Id, Dd, Vd, rd, d_, dolp, dolp_, vP, dP_ = \
-        comp(p, pri_p, fd, fv, x, X,
-             pri_s, I, D, V, rv, p_, olp, olp_,
-             pri_sd, Id, Dd, Vd, rd, d_, dolp, dolp_,
-             a, aV, aD, min_r, A, AV, AD, r, vP_, dP_)
+    else:  # comparison derivation incr within e_ = d_ (not tuples within range incr?)
 
-    return vP_, dP_  # local vPs and dPs to replace p_, A, AV, AD accumulated per comp recursion
+        pri_d = e_[0]  # no deriv_incr while r < min_r, only more fuzzy?
+        fd, fv = 0, 0
 
+        for x in range(1, X):
+            d = e_[x]
 
-def inc_der(a, aV, aD, min_r, A, AV, AD, r, d_):
+            fd, fv, vP, dP, vP_, dP_, olp = \
+            re_comp(x, d, pri_d, fd, fv, vP, dP, vP_, dP_, olp, X, A, r)
 
-    if r > min_r:
-        A += a; AV += aV
-    if r > min_r-1:
-        AD += aD
+            pri_d = d
 
-    X = len(d_)
-    ip_ = d_  # to differentiate from new d_
-
-    fd, fv, r, vP_, dP_ = 0, 0, 0, [], []  # r is initialized for each d_
-    pri_s, I, D, V, rv, olp, p_, olp_ = 0, 0, 0, 0, 0, 0, [], []  # tuple vP=0,
-    pri_sd, Id, Dd, Vd, rd, dolp, d_, dolp_ = 0, 0, 0, 0, 0, 0, [], []  # tuple dP=0
-
-    pri_p = ip_[0]
-
-    for x in range(1, X):
-
-        p = ip_[x]  # better than pop()?
-
-        pri_s, I, D, V, rv, p_, olp, olp_, pri_sd, Id, Dd, Vd, rd, d_, dolp, dolp_, vP, dP_ = \
-        comp(p, pri_p, fd, fv, x, X,
-             pri_s, I, D, V, rv, p_, olp, olp_,
-             pri_sd, Id, Dd, Vd, rd, d_, dolp, dolp_,
-             a, aV, aD, min_r, A, AV, AD, r, vP_, dP_)
-
-        pri_p = p
-
-    return vP_, dP_  # local vPs and dPs to replace d_
+    return vP_, dP_  # local vP_ + dP_ replaces t_ or d_
 
 
-def comp(p, pri_p, fd, fv, x, X,  # input variables
-         pri_s, I, D, V, rv, p_, olp, olp_,  # variables of vP
-         pri_sd, Id, Dd, Vd, rd, d_, dolp, dolp_,  # variables of dP
-         a, aV, aD, min_r, A, AV, AD, r, vP_, dP_):  # filter variables and output patterns
+def form_P(typ, P, alt_P, P_, alt_P_, olp, pri_p, fd, fv, x, X, A, r):
+
+    if typ: s = 1 if fv >= 0 else 0  # sign of fd, 0 is positive?
+    else:   s = 1 if fd >= 0 else 0  # sign of fv, 0 is positive?
+
+    pri_s, I, D, V, rf, e_, olp_ = P  # debug: 0 values in P?
+
+    if x > r + 2 and (s != pri_s or x == X - 1):  # P is terminated and evaluated
+
+        if typ:
+            if len(e_) > r + 3 and pri_s == 1 and V > A + aV:  # minimum of 3 tuples
+                rf = 1  # incr range flag
+                e_.append(pre_comp(1, e_, A, r))  # comparison range incr within e_ = t_
+
+        else:
+            if len(e_) > 3 and abs(D) > A + aD:  # minimum of 3 ds
+                rf = 1  # incr deriv flag
+                r = 1  # consecutive-d comp
+                e_.append(pre_comp(0, e_, A, r))  # comp derivation incr within e_ = d_
+
+        P = pri_s, I, D, V, rf, e_, olp_
+        P_.append(P)  # output to level_2
+
+        o = len(P_), olp  # index of current P and terminated olp are buffered in alt_olp_
+        alt_P[6].append(o)
+        o = len(alt_P_), olp  # index of current alt_P and terminated olp buffered in olp_
+        olp_.append(o)
+
+        olp, I, D, V, rf, e_, olp_ = 0, 0, 0, 0, 0, [], []  # initialized P and olp
+
+    pri_s = s   # vP (span of pixels forming same-sign v) is incremented:
+    I += pri_p  # ps summed within vP
+    D += fd     # fuzzy ds summed within vP
+    V += fv     # fuzzy vs summed within vP
+
+    if typ:
+        t = pri_p, fd, fv  # inputs for inc_rng comp are tuples, vs. pixels for initial comp
+        e_.append(t)
+    else:
+        e_.append(fd)  # prior fds of the same sign are buffered within dP
+
+    P = pri_s, I, D, V, rf, e_, olp_
+
+    return P, alt_P, P_, alt_P_, olp  # alt_ and _alt_ accumulated per level
+
+
+def re_comp(x, p, pri_p, fd, fv, vP, dP, vP_, dP_, olp, X, A, r):  # recursive comp
 
     d = p - pri_p      # difference between consecutive pixels
     m = min(p, pri_p)  # match between consecutive pixels
     v = m - A          # relative match (predictive value) between consecutive pixels
 
-    fd += d  # fd includes all shorter + current- range ds between comparands
-    fv += v  # fv includes all shorter + current- range vs between comparands
+    fd += d  # fuzzy d accumulates ds between p and all prior ps in r via range_incr()
+    fv += v  # fuzzy v; lower-r fv and fd are in lower Ps, different for p and pri_p
+
+    # formation of value pattern vP: span of pixels forming same-sign fv s:
+
+    vP, dP, vP_, dP_, olp = \
+    form_P(1, vP, dP, vP_, dP_, olp, pri_p, fd, fv, x, X, A, r)
+
+    # formation of difference pattern dP: span of pixels forming same-sign fd s:
+
+    dP, vP, dP_, vP_, olp = \
+    form_P(0, dP, vP, dP_, vP_, olp, pri_p, fd, fv, x, X, A, r)
+
+    olp += 1  # overlap between concurrent vP and dP, to be buffered in olp_s
+
+    return fd, fv, vP, dP, vP_, dP_, olp  # for next-p comp, vP and dP increment, output
 
 
-    # formation of value pattern vP: span of pixels forming same-sign v s:
+def comp(x, p, it_, vP, dP, vP_, dP_, olp, X, A, r):  # comparison of pixel to prior r-pixels
 
-    s = 1 if v > 0 else 0  # s: positive sign of v
-    if x > r+2 and (s != pri_s or x == X-1):  # if derived pri_s miss, vP is terminated
+    '''
+    for index in range(0, len(it_)-1):  # accumulation here is not quite right
+        pri_p, fd, fm = it_[index]
 
-        if len(p_) > r+3 and pri_s == 1 and V > AV:  # min 3 comp over extended distance within p_:
+        d = p - pri_p  # difference between pixels
+        m = min(p, pri_p)  # match between pixels
 
-            r += 1  # r: incremental range-of-comp counter
-            rv = 1  # rv: incremental range flag:
-            p_.append(inc_rng(a, aV, aD, min_r, A, AV, AD, r, p_))
+        fd += d  # fuzzy d: sum of ds between p and all prior ps within it_
+        fm += m  # fuzzy m: sum of ms between p and all prior ps within it_
 
-        p = I / len(p_); d = D / len(p_); v = V / len(p_)  # default to eval overlap, poss. div.comp?
-        vP = pri_s, p, I, d, D, v, V, rv, p_, olp_
-        vP_.append(vP)  # output of vP, related to dP_ by overlap only, no discont comp till Le3?
+        it = pri_p, fd, fm
+        it_[index] = it '''
 
-        o = len(vP_), olp  # len(P_) is index of current vP
-        dolp_.append(o)  # indexes of overlapping vPs and olp are buffered at current dP
+    index = 0
 
-        I, D, V, rv, olp, dolp, p_, olp_ = 0, 0, 0, 0, 0, 0, [], []  # initialization of new vP and olp
+    for it in it_:  # incomplete tuples with fd, fm summation range from 0 to r
+        pri_p, fd, fm = it
 
-    pri_s = s   # vP (span of pixels forming same-sign v) is incremented:
-    olp += 1    # overlap to current dP
-    I += pri_p  # ps summed within vP
-    D += fd     # fds summed within vP into fuzzy D
-    V += fv     # fvs summed within vP into fuzzy V
-    pri = pri_p, fd, fv
-    p_.append(pri)  # buffered within vP for selective extended comp
+        d = p - pri_p  # difference between pixels
+        m = min(p, pri_p)  # match between pixels
 
+        fd += d  # fuzzy d: sum of ds between p and all prior ps within it_
+        fm += m  # fuzzy m: sum of ms between p and all prior ps within it_
 
-    # formation of difference pattern dP: span of pixels forming same-sign d s:
+        it = pri_p, fd, fm
+        it_[index] = it
+        index += 1
 
-    sd = 1 if d > 0 else 0  # sd: positive sign of d;
-    if x > r+2 and (sd != pri_sd or x == X-1):  # if derived pri_sd miss, dP is terminated
+    if len(it_) == r:  # current tuple fd and fm are accumulated over range = r
+        fv = fm - A
 
-        if len(d_) > 3 and abs(Dd) > AD:  # min 3 comp within d_:
+        #  formation of value pattern vP: span of pixels forming same-sign fv s:
 
-            rd = 1  # rd: incremental derivation flag:
-            d_.append(inc_der(a, aV, aD, min_r, A, AV, AD, r, d_))
+        vP, dP, vP_, dP_, olp = \
+        form_P(1, vP, dP, vP_, dP_, olp, pri_p, fd, fv, x, X, A, r)
 
-        pd = Id / len(d_); dd = Dd / len(d_); vd = Vd / len(d_)  # so all olp Ps can be directly evaluated
-        dP = pri_sd, pd, Id, dd, Dd, vd, Vd, rd, d_, dolp_
-        dP_.append(dP)  # output of dP
+        # formation of difference pattern dP: span of pixels forming same-sign fd s:
 
-        o = len(dP_), dolp  # len(P_) is index of current dP
-        olp_.append(o)  # indexes of overlapping dPs and dolps are buffered at current vP
+        dP, vP, dP_, vP_, olp = \
+        form_P(0, dP, vP, dP_, vP_, olp, pri_p, fd, fv, x, X, A, r)
 
-        Id, Dd, Vd, rd, olp, dolp, d_, dolp_ = 0, 0, 0, 0, 0, 0, [], []  # initialization of new dP and olp
+        olp += 1  # overlap between vP and dP, stored in both and terminated with either
 
-    pri_sd = sd  # dP (span of pixels forming same-sign d) is incremented:
-    dolp += 1    # overlap to current vP
-    Id += pri_p  # ps summed within dP
-    Dd += fd     # fds summed within dP
-    Vd += fv     # fvs summed within dP
-    d_.append(fd)  # prior fds are buffered within dP, all of the same sign
+    it = p, 0, 0  # fd and fm are directional: initialized at 0 for a new tuple
+    it_.appendleft(it)  # new tuple is added, displacing completed tuple
 
-    return pri_s, I, D, V, rv, p_, olp, olp_, pri_sd, Id, Dd, Vd, rd, d_, dolp, dolp_, vP_, dP_
-    # for next p comparison, vP and dP increment, and output
+    return it_, vP, dP, vP_, dP_, olp  # for next-p comparison, vP and dP increment, output
 
 
-def root_1D(Fp_): # last '_' distinguishes array name from element name
+def root_1D(Fp_):  # last '_' distinguishes array name from element name
 
-    FP_ = []  # output frame of vPs: relative match patterns, and dPs: difference patterns
+    FP_ = []  # output frame of vPs: relative-match patterns, and dPs: difference patterns
     Y, X = Fp_.shape  # Y: frame height, X: frame width
+    min_r = 3  # fuzzy comp range
 
-    a = 127  # minimal filter for vP inclusion
-    aV = 63  # minimal filter for incremental-range comp
-    aD = 63  # minimal filter for incremental-derivation comp
-    min_r=0  # default range of fuzzy comparison, initially 0
+    global a; a = 63
+    global aV; aV = 63 * min_r  # min V for initial incremental-range comp(t_)
+    global aD; aD = 63 * min_r  # min |D| for initial incremental-derivation comp(d_)
+
+    A = a * min_r  # initial min match for positive vP inclusion, += a per recursion
 
     for y in range(Y):
+        p_ = Fp_[y, :]   # y is index of new line p_
 
-        ip_ = Fp_[y, :]  # y is index of new line ip_
+        r, x, olp, vP_, dP_ = min_r, 0, 0, [], []  # initialized at each level
+        vP = 0, 0, 0, 0, 0, [], []  # pri_s, I, D, V, rv, t_, olp_
+        dP = 0, 0, 0, 0, 0, [], []  # pri_sd, Id, Dd, Vd, rd, d_, dolp_
 
-        if min_r == 0: A = a; AV = aV  # actual filters, incremented per comp recursion
-        else: A = 0; AV = 0  # if r > min_r
-
-        if min_r <= 1: AD = aD
-        else: AD = 0
-
-        fd, fv, r, x, vP_, dP_ = 0, 0, 0, 0, [], []  # i/o tuple
-        pri_s, I, D, V, rv, olp, p_, olp_ = 0, 0, 0, 0, 0, 0, [], []  # vP tuple
-        pri_sd, Id, Dd, Vd, rd, dolp, d_, dolp_ = 0, 0, 0, 0, 0, 0, [], []  # dP tuple
-
-        pri_p = ip_[0]
+        it_ = deque(maxlen=r)  # incomplete fuzzy tuples: summation range < r
+        pri_t = p_[0], 0, 0  # no d, m at x = 0
+        it_.append(pri_t)
 
         for x in range(1, X):  # cross-compares consecutive pixels
+            p = p_[x]  # new pixel, fuzzy comp to it_:
 
-            p = ip_[x]  # new pixel for comp to prior pixel, could use pop()?
+            it_, vP, dP, vP_, dP_, olp = \
+            comp(x, p, it_, vP, dP, vP_, dP_, olp, X, A, r)
 
-            pri_s, I, D, V, rv, p_, olp, olp_, pri_sd, Id, Dd, Vd, rd, d_, dolp, dolp_, vP_, dP_ = \
-            comp(p, pri_p, fd, fv, x, X,
-                 pri_s, I, D, V, rv, p_, olp, olp_,
-                 pri_sd, Id, Dd, Vd, rd, d_, dolp, dolp_,
-                 a, aV, aD, min_r, A, AV, AD, r, vP_, dP_)
-
-            pri_p = p  # prior pixel, pri_ values are always derived before use
-
-        LP_ = vP_, dP_
+        LP_ = vP_, dP_   # line of patterns formed from a line of pixels
         FP_.append(LP_)  # line of patterns is added to frame of patterns, y = len(FP_)
 
     return FP_  # output to level 2
@@ -194,3 +210,4 @@ root_1D(f)
 
 # at vP term: print ('type', 0, 'pri_s', pri_s, 'I', I, 'D', D, 'V', V, 'rv', rv, 'p_', p_)
 # at dP term: print ('type', 1, 'pri_sd', pri_sd, 'Id', Id, 'Dd', Dd, 'Vd', Vd, 'rd', rd, 'd_', d_)
+
